Move useConfig
