+++ packages/core/src/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './useBlockMeta'
 export * from './useBlockNumber'
 export * from './useBlockNumbers'
 export * from './useChainCalls'
+export * from './useConfig'
 export * from './useDebounce'
 export * from './useDebouncePair'
 export * from './useEthers'
